+++ packages/moonstone/Scrollable/Scrollable.js
@@ -10,7 +10,7 @@ import classNames from 'classnames';
 import handle, {forward} from '@enact/core/handle';
 import platform from '@enact/core/platform';
 import {onWindowReady} from '@enact/core/snapshot';
-import {Job} from '@enact/core/util';
+import {clamp, Job} from '@enact/core/util';
 import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
 import {constants, ScrollableBase as UiScrollableBase} from '@enact/ui/Scrollable';
 import Spotlight, {getDirection} from '@enact/spotlight';
@@ -528,9 +528,10 @@ class ScrollableBase extends Component {
 				// Should do nothing when focusedItem is paging control button of Scrollbar
 				if (childRefCurrent.containerRef.current.contains(focusedItem)) {
 					const
+						contentRect = this.uiRef.current.childRefCurrent.containerRef.current.getBoundingClientRect(),
 						clientRect = focusedItem.getBoundingClientRect(),
-						x = (clientRect.right + clientRect.left) / 2,
-						y = (clientRect.bottom + clientRect.top) / 2;
+						x = clamp(contentRect.left, contentRect.right, (clientRect.right + clientRect.left) / 2),
+						y = clamp(contentRect.top, contentRect.bottom, (clientRect.bottom + clientRect.top) / 2);
 
 					focusedItem.blur();
 					if (!this.props['data-spotlight-container-disabled']) {
@@ -712,13 +713,15 @@ class ScrollableBase extends Component {
 				const position = {x, y};
 				const {current: {containerRef: {current}}} = this.uiRef;
 				const elemFromPoint = document.elementFromPoint(x, y);
-				const target =
-					getIntersectingElement(elemFromPoint.closest(`.${spottableClass}`), current) ||
-					getTargetInViewByDirectionFromPosition(direction, position, current) ||
-					getTargetInViewByDirectionFromPosition(reverseDirections[direction], position, current);
+				if (elemFromPoint) {
+					const target =
+						getIntersectingElement(elemFromPoint.closest(`.${spottableClass}`), current) ||
+						getTargetInViewByDirectionFromPosition(direction, position, current) ||
+						getTargetInViewByDirectionFromPosition(reverseDirections[direction], position, current);
 
-				if (target) {
-					Spotlight.focus(target);
+					if (target) {
+						Spotlight.focus(target);
+					}
 				}
 			}
 			this.pointToFocus = null;