+++ packages/moonstone/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
-- `moonstone/VirtualList.VirtualGridList` and `moonstone/VirtualList.VirtualList` to restore focus properly when pressing page up after holding 5-way down
+- `moonstone/Scroller` to restore focus properly when pressing page up after holding 5-way down
 
 ## [3.0.0] - 2019-09-03